# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  handlers.UpdateUserRequest:
    properties:
      email_verified:
        type: boolean
      role:
        enum:
        - user
        - admin
        - superadmin
        type: string
      subscription_tier:
        enum:
        - free
        - starter
        - pro
        - business
        type: string
      usage_limit:
        minimum: 0
        type: integer
    type: object
  importpkg.ImportError:
    properties:
      field:
        type: string
      message:
        type: string
      row:
        type: integer
      value:
        type: string
    type: object
  importpkg.ImportResult:
    properties:
      duration:
        type: string
      errors:
        items:
          $ref: '#/definitions/importpkg.ImportError'
        type: array
      failure_count:
        type: integer
      imported_leads:
        items:
          $ref: '#/definitions/importpkg.ImportedLead'
        type: array
      success_count:
        type: integer
      total_rows:
        type: integer
    type: object
  importpkg.ImportedLead:
    properties:
      industry:
        type: string
      lead_id:
        type: string
      name:
        type: string
      row:
        type: integer
    type: object
...
  models.ErrorResponse:
    properties:
      error:
        type: string
      message:
        type: string
    type: object
...
  models.UserResponse:
    properties:
      created_at:
        type: string
      email:
        type: string
      email_verified:
        type: boolean
      id:
        type: integer
      name:
        type: string
      role:
        type: string
      subscription_tier:
        type: string
      usage_count:
        type: integer
      usage_limit:
        type: integer
    type: object
host: localhost:7890
info:
  contact:
    email: support@industrydb.io
    name: API Support
    url: https://industrydb.io/support
  description: Industry-specific business data API. Verified. Affordable.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://industrydb.io/terms
  title: IndustryDB API
  version: "1.0"
paths:
...
  /admin/import/csv:
    post:
      consumes:
      - multipart/form-data
      description: Bulk import leads from CSV file (admin only) - max 10k rows per
        upload
      parameters:
      - description: CSV file to import
        in: formData
        name: file
        required: true
        type: file
      - description: Only validate, don't import
        in: formData
        name: validate_only
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: Import results
          schema:
            $ref: '#/definitions/importpkg.ImportResult'
        "400":
          description: Invalid file or format
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Forbidden - Admin access required
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "413":
          description: File too large
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Import leads from CSV
      tags:
      - Admin
  /admin/stats:
    get:
      description: Get aggregated statistics about users, subscriptions, and exports
        (admin only)
      produces:
      - application/json
      responses:
        "200":
          description: Platform statistics
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Forbidden - Admin access required
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get platform statistics
      tags:
      - Admin
  /admin/users:
    get:
      description: Get paginated list of users with optional filters (admin only)
      parameters:
      - description: 'Page number (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Items per page (default: 50, max: 100)'
        in: query
        name: limit
        type: integer
      - description: Filter by subscription tier (free, starter, pro, business)
        in: query
        name: tier
        type: string
      - description: Filter by email verification (true, false)
        in: query
        name: verified
        type: string
      - description: Filter by user role (user, admin, superadmin)
        in: query
        name: role
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of users with pagination
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Forbidden - Admin access required
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List all users
      tags:
      - Admin
  /admin/users/{id}:
    delete:
      description: Suspend (soft delete) a user account - cannot suspend yourself
        or superadmins (admin only)
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: User suspended successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Cannot suspend own account or superadmin
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Forbidden - Admin access required
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Suspend user account
      tags:
      - Admin
    get:
      description: Get detailed information about a specific user including subscriptions,
        exports, and recent audit logs (admin only)
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: User details
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid user ID
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Forbidden - Admin access required
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get user details
      tags:
      - Admin
    patch:
      consumes:
      - application/json
      description: Update user subscription tier, role, email verification status,
        or usage limit (admin only)
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: User update data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Updated user
          schema:
            $ref: '#/definitions/models.UserResponse'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Forbidden - Admin access required
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update user
      tags:
      - Admin
...
securityDefinitions:
  ApiKeyAuth:
    description: API Key for programmatic access (Business tier only)
    in: header
    name: X-API-Key
    type: apiKey
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"